rename dragdropgame file
